+++ requests_cache/backends/sqlite.py
@@ -6,7 +6,7 @@ from os import makedirs
 from os.path import abspath, basename, dirname, expanduser, isabs, join
 from pathlib import Path
 from tempfile import gettempdir
-from typing import Iterable, List, Tuple, Type, Union
+from typing import Collection, Iterable, Iterator, List, Tuple, Type, Union
 
 from . import BaseCache, BaseStorage, get_valid_kwargs
 
@@ -87,7 +87,7 @@ class DbDict(BaseStorage):
         connection.execute(f'CREATE TABLE IF NOT EXISTS {self.table_name} (key PRIMARY KEY, value)')
 
     @contextmanager
-    def connection(self, commit=False) -> sqlite3.Connection:
+    def connection(self, commit=False) -> Iterator[sqlite3.Connection]:
         """Get a thread-local database connection"""
         if not hasattr(self._local_context, 'con'):
             logger.debug(f'Opening connection to {self.db_path}:{self.table_name}')
@@ -191,7 +191,7 @@ class DbPickleDict(DbDict):
         return self.serializer.loads(super().__getitem__(key))
 
 
-def _format_sequence(values: Iterable) -> Tuple[str, List]:
+def _format_sequence(values: Collection) -> Tuple[str, List]:
     """Get SQL parameter marks for a sequence-based query, and ensure value is a sequence"""
     if not isinstance(values, Iterable):
         values = [values]