+++ requests_cache/cache_control.py
@@ -3,7 +3,7 @@ from datetime import datetime, timedelta, timezone
 from email.utils import parsedate_to_datetime
 from fnmatch import fnmatch
 from logging import getLogger
-from typing import Any, Dict, Optional, Tuple, Union
+from typing import Any, Dict, Mapping, Optional, Tuple, Union
 
 from requests import PreparedRequest, Response
 
@@ -55,7 +55,7 @@ class CacheActions:
         else:
             self._init_from_settings(url=request.url, **kwargs)
 
-    def _init_from_headers(self, headers: Dict):
+    def _init_from_headers(self, headers: Mapping):
         """Initialize from request headers"""
         directives = get_cache_directives(headers)
         do_not_cache = directives.get('max-age') == DO_NOT_CACHE
@@ -122,7 +122,7 @@ def get_expiration_datetime(expire_after: ExpirationTime) -> Optional[datetime]:
     return datetime.utcnow() + expire_after
 
 
-def get_cache_directives(headers: Dict) -> Dict:
+def get_cache_directives(headers: Mapping) -> Dict:
     """Get all Cache-Control directives, and handle multiple headers and comma-separated lists"""
     if not headers:
         return {}
@@ -135,8 +135,13 @@ def get_cache_directives(headers: Dict) -> Dict:
     return kv_directives
 
 
-def get_url_expiration(url: str, urls_expire_after: ExpirationPatterns = None) -> ExpirationTime:
+def get_url_expiration(
+    url: Optional[str], urls_expire_after: ExpirationPatterns = None
+) -> ExpirationTime:
     """Check for a matching per-URL expiration, if any"""
+    if not url:
+        return None
+
     for pattern, expire_after in (urls_expire_after or {}).items():
         if url_match(url, pattern):
             logger.debug(f'URL {url} matched pattern "{pattern}": {expire_after}')
@@ -144,7 +149,7 @@ def get_url_expiration(url: str, urls_expire_after: ExpirationPatterns = None) -
     return None
 
 
-def has_cache_headers(headers: Dict) -> bool:
+def has_cache_headers(headers: Mapping) -> bool:
     """Determine if headers contain cache directives **that we currently support**"""
     has_cache_control = any([d in headers.get('Cache-Control', '') for d in CACHE_DIRECTIVES])
     return has_cache_control or bool(headers.get('Expires'))